deepen: fold strict-schema converter into its only consumer (the test)
Deletes lib/ai/prompts/strict-schema.ts (129 LOC). The module exported
toStrictSchema and auditStrictCompliance; both were imported only by
__tests__/lib/ai/strict-schema-audit.test.ts. No production code ever
called toStrictSchema — production schemas in unified-prompts.ts flow
straight to summarize.ts without conversion, and the OpenRouter client
does not pass response_format to xAI Grok today.

Per LANGUAGE.md, one adapter is a hypothetical seam, not a real one.
This module was test infrastructure shaped like production code under
lib/. Folding it into the test concentrates locality: the strict-mode
wire-shape contract and its compliance walker now sit beside the
assertions that exercise them.

The schemas themselves still live behind their existing deep interface
(CONTEXT.md "Filing Prompt" — getSchemaForFormType in unified-prompts.ts).
This change touches neither the production interface nor the schemas.

Updated the GROUNDING_BLOCK comment in unified-prompts.ts that
previously referenced "the strict-schema converter" — it now points to
the test that documents the strict-mode wire-shape contract, so a future
reader looking up the reference lands at live code rather than a deleted
module.

If response_format strict mode is ever turned on at the OpenRouter seam
(adding a second adapter), re-extract the converter into a real module.
Until then, the hypothetical seam stays inlined with its consumer.

Tests deleted: 0 (the test moves intact; same 52 assertions pass)
Tests added: 0
Test surface: __tests__/lib/ai/strict-schema-audit.test.ts unchanged
in behaviour — 52 tests, all green.
